+++ packages/revealui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # revealui
 
+## 0.1.14
+
+### Patch Changes
+
+- create-revealui@0.5.19
+
 ## 0.1.13
 
 ### Patch Changes